+++ packages/@haiku/core/src/adapters/dom/HaikuDOMAdapter.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Haiku 2016-2018. All rights reserved.
  */
 
-import {BytecodeOptions, HaikuBytecode} from '../../api/HaikuBytecode';
+import {BytecodeOptions, HaikuBytecode} from '../../api';
 import HaikuContext, {ComponentFactory} from './../../HaikuContext';
 import dom from './../../renderers/dom';
