--- packages/@haiku/core/src/reflection/arrayToRO.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) Haiku 2016-2018. All rights reserved.
- */
-
-import expressionToRO from './expressionToRO';
-
-export default function arrayToRO (arr) {
-  const out = [];
-  for (let i = 0; i < arr.length; i++) {
-    out[i] = expressionToRO(arr[i], null);
-  }
-  return out;
-}
