--- packages/@haiku/core/src/reflection/objectToRO.ts
@@ -1,16 +1,0 @@
-/**
- * Copyright (c) Haiku 2016-2018. All rights reserved.
- */
-
-import expressionToRO from './expressionToRO';
-
-export default function objectToRO (obj, options) {
-  const out = {};
-  for (const key in obj) {
-    if (options && options.ignore && options.ignore.test(key)) {
-      continue;
-    }
-    out[key] = expressionToRO(obj[key], options);
-  }
-  return out;
-}
