+++ packages/@haiku/core/src/renderers/dom/HaikuDOMRenderer.ts
@@ -4,6 +4,7 @@
 
 import HaikuBase, {GLOBAL_LISTENER_KEY} from './../../HaikuBase';
 import HaikuComponent from './../../HaikuComponent';
+import assign from './../../vendor/assign';
 import applyLayout from './applyLayout';
 import assignAttributes from './assignAttributes';
 import cloneVirtualElement from './cloneVirtualElement';
@@ -16,7 +17,6 @@ import getLocalDomEventPosition from './getLocalDomEventPosition';
 import getTypeAsString from './getTypeAsString';
 import isBlankString from './isBlankString';
 import isTextNode from './isTextNode';
-import mixpanelInit from './mixpanelInit';
 import normalizeName from './normalizeName';
 import removeElement from './removeElement';
 import replaceElementWithText from './replaceElementWithText';
@@ -843,4 +843,46 @@ const SVG_ELEMENT_NAMES = {
   use: true,
   view: true,
   vkern: true,
+};
+
+/* tslint:disable */
+const mixpanelSnippetInjector = new Function(`
+  (function(e,a){if(!a.__SV){var b=window;try{var c,l,i,j=b.location,g=j.hash;c=function(a,b){return(l=a.match(RegExp(b+"=([^&]*)")))?l[1]:null};g&&c(g,"state")&&(i=JSON.parse(decodeURIComponent(c(g,"state"))),"mpeditor"===i.action&&(b.sessionStorage.setItem("_mpcehash",g),history.replaceState(i.desiredHash||"",e.title,j.pathname+j.search)))}catch(m){}var k,h;window['mixpanel']=a;a._i=[];a.init=function(b,c,f){function e(b,a){var c=a.split(".");2==c.length&&(b=b[c[0]],a=c[1]);b[a]=function(){b.push([a].concat(Array.prototype.slice.call(arguments,
+  0)))}}var d=a;"undefined"!==typeof f?d=a[f]=[]:f="mixpanel";d.people=d.people||[];d.toString=function(b){var a="mixpanel";"mixpanel"!==f&&(a+="."+f);b||(a+=" (stub)");return a};d.people.toString=function(){return d.toString(1)+".people (stub)"};k="disable time_event track track_pageview track_links track_forms register register_once alias unregister identify name_tag set_config reset people.set people.set_once people.unset people.increment people.append people.union people.track_charge people.clear_charges people.delete_user".split(" ");
+  for(h=0;h<k.length;h++)e(d,k[h]);a._i.push([b,c,f])};a.__SV=1.2;}})(document,window.mixpanel||[]);
+`);
+/* tslint:enable */
+
+declare var window: any;
+
+const mixpanelInit = (mixpanelToken, component) => {
+  // Only initialize Mixpanel if we're running in the browser
+  if (typeof window !== 'undefined') {
+    // Don't initialize multiple times if multiple components are on the page
+    if (!window.mixpanel) {
+      mixpanelSnippetInjector();
+
+      const head = document.getElementsByTagName('head')[0];
+
+      const script = document.createElement('script');
+      script.async = true;
+      script.type = 'text/javascript';
+      script.src = 'https://cdn.mxpnl.com/libs/mixpanel-2-latest.min.js';
+      head.appendChild(script);
+
+      window.mixpanel.init(mixpanelToken, {ip: false});
+    }
+
+    const metadata = (component.bytecode && component.bytecode.metadata) || {};
+
+    window.mixpanel.track(
+      'component:initialize',
+      assign(
+        {
+          platform: 'dom',
+        },
+        metadata,
+      ),
+    );
+  }
 };
